refactor: move stack implementation to internal folder
